add value priorities metric toggle
